Add a locomotion system
It is possible to move using the knob on the pad. Moving in the real
world affects the character in the game. Crouching allows to pass
through lower passages as well as slows down the player in the game.
Sprinting and jumping have been added. Some objects do not collide with
the player when grabbing them in the hand. Camera snapping also works
properly.