Add first demo level, fix and extend portals, add narrator
I added a NarratorBox that allows to activate sound in a particular
location. The new PortalVR_Extended allows to react with sound / target
portal changes or sending notifications to any selected portal when
players enters it for n-th time (for example, second, third, or from 50
to 100). Portals now activate when the player is in the designated
trigger box - improving performance. Fixed a bug where the portal was
teleporting to random locations or activating too many times in one go.
And much more...